Refactor JVM DTO renderers
